yeah

+++ src/endpoints/submissions.ts
@@ -4,8 +4,9 @@ import rounds from "../rounds.json"
 import { modifiersToSubmissions, submissions, users, usersToSubmissions, modifiers, requests } from '../db/schema';
 import { DB } from "../modules/db";
 import snowflake from "../modules/snowflake";
-import { eq } from "drizzle-orm";
+import { and, eq } from "drizzle-orm";
 import { makeRequest } from "../modules/requests";
+import { getRoundsObj } from "./rounds";
 
 Pointer.GET(AuthLevels.ONLY_DISCORD, `/submit/:round`, async (req, c, pack) => {
   if (pack.user == undefined) { return new Error("User Invalid!") }
@@ -42,9 +43,8 @@ Pointer.GET(AuthLevels.ONLY_DISCORD, `/submit/:round`, async (req, c, pack) => {
       challenger: undefined,
       title: (submission.title as string),
       link: (submission.link as string),
-      player_link: (submission.player_link as string),
       desc: (submission.desc as string),
-      round_id: round,
+      round: round,
       modifiers: submission.modifiers.map(entry => entry.modifierId),
       okay_count: 0, // Empty here
       score: 0 // Empty here
@@ -58,6 +58,9 @@ Pointer.GET(AuthLevels.ONLY_DISCORD, `/submit/:round`, async (req, c, pack) => {
 Pointer.POST(AuthLevels.ONLY_DISCORD, `/submit/:round`, async (req, c, pack) => {
   // Check User
   if (pack.user == undefined) { return new Error("User Invalid!") }
+
+  // Is Participant
+  if (!pack.user.participant) { return new Error("User Not A Participant!") } // <- I can't believe you never checked that
 
   // Get and Check Round
   let raw_round = req.param("round")
@@ -77,15 +80,30 @@ Pointer.POST(AuthLevels.ONLY_DISCORD, `/submit/:round`, async (req, c, pack) => 
   
   if (round > id) { return new Error("Round Not Even Started Yet...") }
 
+  // Check if Round is Still in Open Phase!
+  let current_round = getRoundsObj(round)
+  if (current_round == undefined) { return new Error("Round Invalid!") }
+  if (Date.now() > current_round.vote) { return new Error("Round No Longer Accepting Submissions!") }
+
   // DB Worker
   const db = DB(c)
 
-  // Get Previous Submission For Round to Replace
+  // new ID define
+  let newID: string = ""
+
+  // Get Previous Submission For Round to Replace OR update
+  let updating = false
+  let keeping = false
   let user = await db.query.users.findFirst({
     where: eq(users.id, pack.user.id),
     with: {
       submissions: {with: {
-        submission: true
+        submission: {
+          with: {
+            authors: true,
+            modifiers: true
+          }
+        }
       }}
     }
   })
@@ -94,7 +112,27 @@ Pointer.POST(AuthLevels.ONLY_DISCORD, `/submit/:round`, async (req, c, pack) => 
   })
   print("problem_submission: ", problem_submission)
   if (problem_submission != null) { // Deleting to Replace...
-    await db.delete(submissions).where(eq(submissions.id, problem_submission.submissionId))
+    if (problem_submission.submission.authors.length > 1) { // UPDATE
+      newID = problem_submission.submissionId
+      updating = true
+    } else if (problem_submission.submission.challengerId != null) { // UPDATE
+      newID = problem_submission.submissionId
+      keeping = true
+    } else {
+      await db.delete(submissions).where(eq(submissions.id, problem_submission.submissionId))
+    }
+  }
+  print("updating or keeping?", updating, keeping)
+  let thistest = !(updating || keeping)
+
+  if (thistest && problem_submission != null) {
+    await db.delete(modifiersToSubmissions).where(eq(modifiersToSubmissions.submissionId, problem_submission.submissionId))
+    // let clearModifierRelationsProms: Promise<any>[] = []
+    // problem_submission.submission.modifiers.forEach(entry => {
+    //   clearModifierRelationsProms.push()
+    // })
+
+    // await Promise.all(clearModifierRelationsProms)
   }
 
   // Get Submission Data
@@ -106,24 +144,35 @@ Pointer.POST(AuthLevels.ONLY_DISCORD, `/submit/:round`, async (req, c, pack) => 
   if (SongURL(submission.link) == null) { return new Error("Submission Link Invalid! (Possibly Disallowed Mod/URL Origin)") }
 
   // Database Submission Data
-  let newSubmission = await db.insert(submissions).values({
-    id: snowflake(),
-    title: submission.title,
-    link: submission.link,
-    desc: submission.desc,
-    round,
-    submitter: pack.user.id
-  }).returning({ id: submissions.id })
+  print("thistest", thistest)
+  if (thistest) {
+    newID = (await db.insert(submissions).values({
+      id: snowflake(),
+      title: submission.title,
+      link: submission.link,
+      desc: submission.desc,
+      round,
+      submitter: pack.user.id
+    }).returning({ id: submissions.id }))[0].id
+  } else {
+    let thisSubmissionObj: any = {...submission}
+    if (keeping) {
+      thisSubmissionObj["challengerId"] = (keeping ? problem_submission?.submission?.challengerId : null)
+    }
+    await db.update(submissions).set(thisSubmissionObj).where(eq(submissions.id, newID))
+  }
 
   // Database Submission Authors
   let userId = pack.user.id
-  let submissionId = (newSubmission[0].id)
+  let submissionId = (newID)
 
   let inserts: Promise<any>[] = []
-  inserts.push(db.insert(usersToSubmissions).values({
-    userId,
-    submissionId
-  }))
+  if (thistest) {
+    inserts.push(db.insert(usersToSubmissions).values({
+      userId,
+      submissionId
+    }))
+  }
 
   // Database Submission Modifiers
   submission.modifiers.forEach((modifierId: string) => {
@@ -134,7 +183,7 @@ Pointer.POST(AuthLevels.ONLY_DISCORD, `/submit/:round`, async (req, c, pack) => 
   })
 
   // Send Out Requests
-  if (submission.request_type != null && submission.request_receivingId != null) {
+  if (!updating && !keeping && submission.request_type != null && submission.request_receivingId != null) {
     inserts.push(makeRequest(
       submission.request_type,
       userId,
@@ -146,5 +195,5 @@ Pointer.POST(AuthLevels.ONLY_DISCORD, `/submit/:round`, async (req, c, pack) => 
 
   await Promise.all(inserts)
 
-  return newSubmission
+  return {id: newID}
 })
